backend/requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
python-dotenv==1.0.0
openai==1.3.0
requests==2.31.0
numpy==1.24.3
faiss-cpu==1.7.4
chromadb==0.4.15
PyPDF2==3.0.1
python-docx==0.8.11
python-multipart==0.0.6
Werkzeug==2.3.7
